refactor: 컴포넌트 컨테이너 id 상수 객체 생성
