+++ vanilla-todo/src/constants/Container.js
@@ -1,0 +1,8 @@
+const CONTAINER = {
+  ADD_FORM: "AddForm",
+  EDIT_FORM: "EditForm",
+  TITLE: "Title",
+  TODO_LIST: "TodoList",
+};
+
+export default CONTAINER;
