+++ vanilla-todo/src/constants/Selector.js
@@ -1,7 +1,9 @@
-export const SELECTOR = {
+const SELECTOR = {
   ADD_FORM_ID: "addForm",
   EDIT_FORM_CLASSNAME: "editForm",
   EDIT_BUTTON_CLASSNAME: "editing",
   TOGGLE_CLASSNAME: "toggle",
   DELETE_BUTTON_CLASSNAME: "delete",
 };
+
+export default SELECTOR;
