+++ vanilla-todo/src/constants/_index.js
@@ -1,0 +1,4 @@
+import SELECTOR from "./Selector.js";
+import CONTAINER from "./Container.js";
+
+export { SELECTOR, CONTAINER };
